# contracts/openapi/itemba-z.v1.yaml
openapi: 3.1.0
info:
  title: ITEMBA-Z Platform API
  version: 1.0.0
  description: >-
    Versioned contract for the ITEMBA-Z Control Center, Sales POS, and approved
    integrations. Posted-sale money uses integer minor units: exactly 100 stored
    minor units equal TZS 1. The current transactional slice and live bootstrap
    read models use whole-unit quantities. Every JSON integer is constrained to
    the JavaScript safe-integer range even though PostgreSQL retains bigint.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /
    description: Current deployment origin
security:
  - oidc: []
...
paths:
...
  /v1/context:
    get:
      summary: Get the caller's working context
      operationId: getWorkingContext
      tags: [Context]
      x-implementation-status: implemented
      responses:
        '200':
          description: Current authorized working context.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkingContext'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /v1/customers:
    get:
      summary: List scoped customers
      operationId: listCustomers
      tags: [Customers]
      x-implementation-status: implemented
      parameters:
        - name: query
          in: query
          schema: { type: string, maxLength: 120 }
        - name: credit_eligible
          in: query
          schema: { type: boolean }
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Customers visible in the caller's scope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerPage'
        '403':
          $ref: '#/components/responses/Forbidden'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/products:
    get:
      summary: List scoped products and availability
      operationId: listProducts
      tags: [Products]
      x-implementation-status: implemented
      parameters:
        - name: query
          in: query
          schema: { type: string, maxLength: 120 }
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Authorized products, prices, and available quantities.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductPage'
        '403':
          $ref: '#/components/responses/Forbidden'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/sales:
    post:
      summary: Complete and atomically post a sale
      operationId: completeSale
      tags: [Sales]
      x-implementation-status: implemented
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompleteSaleCommand'
      responses:
        '201':
          description: Sale completed, or its idempotent original result returned, exactly once.
          headers:
            Location:
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sale'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationFailed'
    get:
      summary: List scoped sales
      operationId: listSales
      tags: [Sales]
      x-implementation-status: implemented
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Sales visible in the caller's ownership and organization scope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalePage'
        '403':
          $ref: '#/components/responses/Forbidden'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/sales/{sale_id}:
    get:
      summary: Get an immutable posted sale
      operationId: getSale
      tags: [Sales]
      x-implementation-status: implemented
      parameters:
        - $ref: '#/components/parameters/SaleId'
      responses:
        '200':
          description: Immutable posted sale.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sale'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/sales/{sale_id}/reversals:
    post:
      summary: Post a linked sale reversal
      operationId: reverseSale
      tags: [Sales]
      x-implementation-status: implemented
      parameters:
        - $ref: '#/components/parameters/SaleId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReverseSaleCommand'
      responses:
        '201':
          description: Linked reversal posted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sale'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationFailed'
  /v1/mobile/sync/sales:
    post:
      summary: Synchronize a mobile sale idempotently
      description: >-
        Synchronizes cash sales from an enrolled device. CREDIT commands are
        rejected with 422 even while online until the mobile workflow carries
        an authoritative AR-aging and approval decision.
      operationId: syncMobileSale
      tags: [Mobile Sync]
      x-implementation-status: implemented
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MobileSaleSyncCommand'
      responses:
        '200':
          description: Existing or newly completed sale result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileSyncResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationFailed'
  /v1/mobile/devices/enroll:
    post:
      summary: Enroll or refresh a POS device binding
      description: >-
        Binds the device to the authenticated actor and exact tenant, company,
        branch, and warehouse scope. Re-enrollment is idempotent only for that
        same binding and returns authoritative offline policy and allocations.
        The acknowledged app and installed version fields are advanced only
        when both optional installed-version acknowledgements equal the current
        available versions. A successful acknowledgement also issues a bounded,
        server-persisted offline-sales lease for that exact app and cache pair.
        A client must atomically install its cache before acknowledging it; a
        repeated acknowledgement is safe after a lost response.
      operationId: enrollMobileDevice
      tags: [Mobile Sync]
      x-implementation-status: implemented
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MobileDeviceEnrollmentCommand'
      responses:
        '200':
          description: Current authoritative device binding and policy.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileDeviceEnrollment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationFailed'
components:
  securitySchemes:
    oidc:
      type: openIdConnect
      openIdConnectUrl: https://identity.itemba-z.example/.well-known/openid-configuration
      description: >-
        Production tokens must contain UUID-valued user_id, tenant_id,
        company_id, branch_id, and warehouse_id claims. The standard sub claim
        remains identity-provider provenance and is not used as a database key.
  parameters:
...
    Cursor:
      name: cursor
      in: query
      schema: { type: string }
    PageSize:
      name: page_size
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 16, maxLength: 128 }
    CorrelationId:
      name: X-Correlation-ID
      in: header
      required: false
      schema: { type: string, format: uuid }
    SaleId:
      name: sale_id
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    BadRequest:
      description: Request syntax or parameters are invalid.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unauthorized:
      description: Authentication is missing or invalid.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: Caller lacks permission or organization scope.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: Resource is absent or not visible to the caller.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Conflict:
      description: Current resource state conflicts with the command.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    ValidationFailed:
      description: Business validation rejected the command.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
...
    WorkingContext:
      type: object
      additionalProperties: false
      required: [actor_id, tenant_id, company_id, company_name, branch_id, branch_name, warehouse_id, warehouse_name, currency, locale, timezone, permissions, master_data_version, price_version]
      properties:
        actor_id: { type: string, format: uuid }
        tenant_id: { type: string, format: uuid }
        company_id: { type: string, format: uuid }
        company_name: { type: string }
        branch_id: { type: string, format: uuid }
        branch_name: { type: string }
        warehouse_id: { type: string, format: uuid }
        warehouse_name: { type: string }
        currency: { type: string, const: TZS }
        locale: { type: string, enum: [en-TZ, sw-TZ] }
        timezone: { type: string, example: Africa/Dar_es_Salaam }
        permissions:
          type: array
          items: { type: string }
        master_data_version: { $ref: '#/components/schemas/SafePositiveInteger' }
        price_version: { $ref: '#/components/schemas/SafePositiveInteger' }
...
    CustomerPage:
      type: object
      additionalProperties: false
      required: [items, next_cursor]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/CustomerSummary' } }
        next_cursor: { type: [string, 'null'] }
...
    ProductPage:
      type: object
      additionalProperties: false
      required: [items, next_cursor]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/ProductSummary' } }
        next_cursor: { type: [string, 'null'] }
...
    SaleKind:
      type: string
      enum: [CASH, CREDIT]
...
    SaleRecordType:
      type: string
      enum: [SALE, REVERSAL]
...
    CompleteSaleCommand:
      type: object
      additionalProperties: false
      required: [customer_id, kind, lines]
      properties:
        customer_id: { type: string, format: uuid }
        kind: { $ref: '#/components/schemas/SaleKind' }
        payment_method: { $ref: '#/components/schemas/PaymentMethod' }
        lines:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/SaleLineCommand' }
...
    Sale:
      type: object
      additionalProperties: false
      required: [id, scope, record_type, kind, status, customer_id, currency, subtotal_minor, tax_minor, total_minor, receipt_reference, fiscal_status, created_by, correlation_id, created_at, lines]
      properties:
        id: { type: string, format: uuid }
        scope: { $ref: '#/components/schemas/Scope' }
        record_type: { $ref: '#/components/schemas/SaleRecordType' }
        kind: { $ref: '#/components/schemas/SaleKind' }
        status: { $ref: '#/components/schemas/SaleStatus' }
        customer_id: { type: string, format: uuid }
        currency: { type: string, pattern: '^[A-Z]{3}$' }
        subtotal_minor: { $ref: '#/components/schemas/SafeNonNegativeInteger' }
        tax_minor: { $ref: '#/components/schemas/SafeNonNegativeInteger' }
        total_minor: { $ref: '#/components/schemas/SafeNonNegativeInteger' }
        payment_method: { $ref: '#/components/schemas/PaymentMethod' }
        device_id: { type: string, format: uuid }
        client_transaction_id: { type: string, format: uuid }
        client_timestamp: { type: string, format: date-time }
        app_version: { type: string, minLength: 1, maxLength: 64 }
        master_data_version: { $ref: '#/components/schemas/SafePositiveInteger' }
        price_version: { $ref: '#/components/schemas/SafePositiveInteger' }
        offline:
          type: boolean
          description: True only for controlled physical-CASH, zero-rated mobile sales in this milestone.
        receipt_reference:
          type: string
          description: Authoritative non-fiscal receipt reference; never a TRA fiscal receipt number.
        fiscal_status: { $ref: '#/components/schemas/FiscalStatus' }
        reversal_of: { type: string, format: uuid }
        reversal_reason: { type: string }
        created_by: { type: string, format: uuid }
        correlation_id: { type: string, format: uuid }
        created_at: { type: string, format: date-time }
        reversed_at: { type: string, format: date-time }
        lines: { type: array, items: { $ref: '#/components/schemas/SaleLine' } }
    SalePage:
      type: object
      additionalProperties: false
      required: [items, next_cursor]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/Sale' } }
        next_cursor: { type: [string, 'null'] }
    ReverseSaleCommand:
      type: object
      additionalProperties: false
      required: [reason]
      properties:
        reason: { type: string, minLength: 1, maxLength: 500 }
...
    MobileSaleSyncCommand:
      type: object
      additionalProperties: false
      required: [customer_id, kind, lines, device_id, client_transaction_id, client_timestamp, app_version, master_data_version, price_version, sync_attempt]
      properties:
        customer_id: { type: string, format: uuid }
        kind: { $ref: '#/components/schemas/SaleKind' }
        payment_method: { $ref: '#/components/schemas/PaymentMethod' }
        lines:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/SaleLineCommand' }
        device_id: { type: string, format: uuid }
        client_transaction_id: { type: string, format: uuid }
        client_timestamp: { type: string, format: date-time }
        app_version: { type: string, minLength: 1, maxLength: 64 }
        master_data_version: { $ref: '#/components/schemas/SafePositiveInteger' }
        price_version: { $ref: '#/components/schemas/SafePositiveInteger' }
        sync_attempt: { type: integer, minimum: 1, maximum: 9007199254740991 }
        offline:
          type: boolean
          default: false
          description: Controlled offline mode accepts only physical CASH and zero-rated product lines. The client timestamp must fall inside a persisted server-issued lease bound to this exact device, app, master-data version, and price version; nonzero effective tax or a changed product price version is rejected with 422.
...
    SafePositiveInteger:
      type: integer
      format: int64
      minimum: 1
      maximum: 9007199254740991
      description: Positive JSON integer constrained to the JavaScript Number safe-integer range.
...